+++ src/query-ast-to-sql-ast/index.js
@@ -523,12 +523,15 @@ function handleUnionSelections(
     // we need to figure out what kind of selection this is
     switch (selection.kind) {
       case 'Field':
+        const alias = selection.alias && selection.alias.value
+
         // has this field been requested once already? GraphQL does not protect against duplicates so we have to check for it
         const existingNode = children.find(
           child =>
             child.fieldName === selection.name.value && child.type === 'table'
+            && (child.alias && child.alias.value) === alias
         )
-        let newNode = new SQLASTNode(sqlASTNode)
+        let newNode = new SQLASTNode(sqlASTNode, { alias })
         if (existingNode) {
           newNode = existingNode
         } else {
@@ -635,12 +638,15 @@ function handleSelections(
     switch (selection.kind) {
       // if its another field, recurse through that
       case 'Field':
+        const alias = selection.alias && selection.alias.value
+
         // has this field been requested once already? GraphQL does not protect against duplicates so we have to check for it
         const existingNode = children.find(
           child =>
             child.fieldName === selection.name.value && child.type === 'table'
+            && (child.alias && child.alias.value) === alias
         )
-        let newNode = new SQLASTNode(sqlASTNode)
+        let newNode = new SQLASTNode(sqlASTNode, { alias })
         if (existingNode) {
           newNode = existingNode
         } else {