+++ docs/CHANGELOG.md
@@ -1,5 +1,32 @@
 ### vNEXT 
 
+**Breaking changes:**
+
+It is no longer guaranteed that a field's value is available under `source[fieldName]` in a custom resolver. Instead, custom resolvers on [non-trivial fields](./warnings.md#non-trivial-fields)  need to use GraphQL's default resolver to get the field value:
+
+```javascript
+import { defaultFieldResolver } from 'graphql'
+
+const User = new GraphQLObjectType({
+  //...
+  fields: () => ({
+    //...
+    following: {
+      // ...
+      resolve: (source, args, context, info) => {
+        const value = defaultFieldResolver(source, args, context, info)
+        return processUsers(value)
+      }
+    }
+  })
+})
+```
+
+See the [docs on custom resolvers](./warnings.md#custom-resolvers) for more details.
+
+#### Fixed
+- [#482](https://github.com/join-monster/join-monster/pull/482): Support different arguments for different aliases
+
 ### v3.3.5 (May 28, 2024)
 #### Fixed
 - [#532](https://github.com/join-monster/join-monster/pull/530): Updates mkdocs to 1.5.3.